+++ page.d.ts
@@ -1,18 +1,16 @@
-import type { Writable } from '@ctx-core/store';
 export interface Page {
-    host: string;
-    path: string;
-    params: params_type;
-    query: query_type;
-    error?: error_type;
+    host: $host_type;
+    path: $path_type;
+    params: $params_type;
+    query: $query_type;
+    error?: $error_type;
 }
+export declare type $host_type = string;
+export declare type $path_type = string;
 export declare type PreloadPage = Page;
 export declare type $page_type = Page;
-export interface page_type extends Writable<$page_type> {
+export interface $params_type extends Record<string, string> {
 }
-export interface params_type extends Record<string, string> {
+export interface $query_type extends Record<string, string | string[]> {
 }
-export interface query_type extends Record<string, string | string[]> {
-}
-export interface error_type extends Error {
-}
+export declare type $error_type = Error | undefined;