+++ sapper_page_b.d.ts
@@ -1,9 +1,0 @@
-import { B } from '@ctx-core/object';
-import { Writable } from '@ctx-core/store';
-export declare const sapper_page_b: sapper_page_b_type;
-export declare type $sapper_page_type = object;
-export declare type $maybe_sapper_page_type = null | $sapper_page_type;
-export interface sapper_page_type extends Writable<$maybe_sapper_page_type> {
-}
-export declare type sapper_page_b_type = B<sapper_page_type>;
-export { sapper_page_b as b__page__sapper };