+++ store.d.ts
@@ -1,4 +1,0 @@
-export * from './sapper_page_b';
-export * from './sapper_path_b';
-export * from './sapper_preloading_b';
-export * from './sapper_session_b';