+++ store.ts
@@ -1,5 +1,0 @@
-// This module is deprecated. use import {} from @ctx-core/sapper instead
-export * from './sapper_page_b'
-export * from './sapper_path_b'
-export * from './sapper_preloading_b'
-export * from './sapper_session_b'